feat: System.Text.Json: dynamic converter supports .NET7.0, .NET8.0
