+++ src/Contrib/Caching/Distributed/Tests/Masa.Contrib.Caching.Distributed.StackExchangeRedis.Tests/_Imports.cs
@@ -7,7 +7,5 @@ global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Options;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using StackExchange.Redis;
-global using System.Dynamic;
 global using System.Reflection;
 global using System.Text.Json;
-global using System.Text.Json.Serialization;
