+++ src/Utils/Extensions/Masa.Utils.Extensions.DotNet/_Imports.cs
@@ -9,5 +9,6 @@ global using System.Dynamic;
 global using System.Globalization;
 global using System.Reflection;
 global using System.Text;
+global using System.Text.Json;
 global using System.Text.Json.Serialization;
 global using System.Text.RegularExpressions;
