+++ src/Utils/Extensions/Tests/Masa.Utils.Extensions.DotNet.Tests/Enumerates/UserClaimType.cs
@@ -1,9 +1,7 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-// ReSharper disable once CheckNamespace
-
-namespace Masa.Contrib.Caching.Distributed.StackExchangeRedis.Tests;
+namespace Masa.Utils.Extensions.DotNet.Tests.Enumerates;
 
 public enum UserClaimType
 {
