+++ src/Utils/Extensions/Tests/Masa.Utils.Extensions.DotNet.Tests/_Imports.cs
@@ -1,7 +1,12 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
+global using Masa.Utils.Extensions.DotNet.Tests.Entities;
+global using Masa.Utils.Extensions.DotNet.Tests.Enumerates;
 global using Masa.Utils.Extensions.DotNet.Tests.Infrastructure;
+global using Microsoft.CSharp.RuntimeBinder;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using System.Dynamic;
 global using System.Text;
 global using System.Text.Json;
+global using System.Text.Json.Serialization;
